requirements.txt
absl-py                 0.12.0
astunparse              1.6.3
bert-serving-client     1.10.0
cachetools              4.2.2
certifi                 2020.12.5
chardet                 4.0.0
cycler                  0.10.0
flatbuffers             1.12
gast                    0.3.3
google-auth             1.30.0
google-auth-oauthlib    0.4.4
google-pasta            0.2.0
grpcio                  1.32.0
grpcio-tools            1.38.0
h5py                    2.10.0
idna                    2.10
importlib-metadata      4.0.1
Keras-Preprocessing     1.1.2
kiwisolver              1.3.1
Markdown                3.3.4
matplotlib              3.3.4
noiseprint2             1.0.0
numpy                   1.19.5
oauthlib                3.1.0
opt-einsum              3.3.0
Pillow                  8.2.0
pip                     21.1.1
pkg-resources           0.0.0
protobuf                3.16.0
pyasn1                  0.4.8
pyasn1-modules          0.2.8
pymilvus                1.1.0
pyparsing               2.4.7
python-dateutil         2.8.1
pyzmq                   22.0.3
redis                   3.5.3
requests                2.25.1
requests-oauthlib       1.3.0
rsa                     4.7.2
setuptools              56.1.0
six                     1.15.0
tensorboard             2.5.0
tensorboard-data-server 0.6.1
tensorboard-plugin-wit  1.8.0
tensorflow              2.4.1
tensorflow-estimator    2.4.0
termcolor               1.1.0
typing-extensions       3.7.4.3
ujson                   4.0.2
urllib3                 1.26.4
Werkzeug                1.0.1
wheel                   0.36.2
wrapt                   1.12.1
zipp                    3.4.1
